requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
click==8.1.7
Django==4.2.16
django-ninja==1.3.0
h11==0.14.0
pydantic==2.9.2
pydantic_core==2.23.4
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.32.0
python-dotenv~=1.0.1
gunicorn==23.0.0
mysqlclient==2.2.4